Update I/O handling for better CI output (#2)
Rich is a really great library for use with CLI tools, but internally it's structured around a terminal. Specifically, the issue is that it requires a width. In the absence of one, it picks 80, which is normally fine, but it makes wide log outputs look really bad and take up more space than they should.

This PR tries to determine if pretty terminal output makes sense, and if it doesn't, it foregoes the use of Rich altogether.